Add call in `emit_type_mismatch_suggestions`
